Create struct Snapshot in seperate module
